chore: organised utils into folders
